separate pvc interfaces and impl
